get file/chunk number

+++ store.h
@@ -15,11 +15,11 @@ void close_database();
 int search_chunk(struct chunk_rec *r);
 int search_chunk_local(struct chunk_rec *r);
 void update_chunk(struct chunk_rec *r);
-int64_t get_chunk_number();
+int get_chunk_number();
 
 int search_file(struct file_rec* r);
 void insert_file(struct file_rec* r);
-int64_t get_file_number();
+int get_file_number();
 
 void init_iterator(char *type);
 void close_iterator(char *type);
